+++ src/steps/resource-manager/compute/index.test.ts
@@ -1,11 +1,6 @@
 import { Disk, VirtualMachine } from '@azure/arm-compute/esm/models';
 import { StorageAccount } from '@azure/arm-storage/esm/models';
-import {
-  Entity,
-  getRawData,
-  MappedRelationship,
-  Relationship,
-} from '@jupiterone/integration-sdk-core';
+import { getRawData, Relationship } from '@jupiterone/integration-sdk-core';
 import {
   buildGalleryImageVersionSourceRelationships,
   buildVirtualMachineDiskRelationships,
@@ -991,7 +986,7 @@ describe('rm-compute-shared-image-version-source-relationships', () => {
       context.jobState.collectedRelationships;
 
     expect(mappedImageSourceRelationships).toHaveLength(1);
-    expect(mappedImageSourceRelationships).toCreateValidRelationshipsToEntities(
+    expect(mappedImageSourceRelationships).toTargetEntities(
       virtualMachineEntities,
     );
   }, 10_000);
@@ -1078,76 +1073,8 @@ describe('rm-compute-virtual-machine-managed-identity-relationships', () => {
     expect(context.jobState.collectedRelationships).toHaveLength(
       systemAssignedIdentityCount + userAssignedIdentityCount,
     );
-    expect(
-      context.jobState.collectedRelationships,
-    ).toCreateValidRelationshipsToEntities(servicePrincipalEntities);
+    expect(context.jobState.collectedRelationships).toTargetEntities(
+      servicePrincipalEntities,
+    );
   });
-});
-
-declare global {
-  // eslint-disable-next-line @typescript-eslint/no-namespace
-  namespace jest {
-    interface Matchers<R> {
-      toCreateValidRelationshipsToEntities(entities: Entity[]): R;
-    }
-  }
-}
-
-expect.extend({
-  toCreateValidRelationshipsToEntities(
-    mappedRelationships: MappedRelationship[],
-    entities: Entity[],
-  ) {
-    for (const mappedRelationship of mappedRelationships) {
-      const _mapping = mappedRelationship._mapping;
-      if (!_mapping) {
-        throw new Error(
-          'expect(mappedRelationships).toCreateValidRelationshipsToEntities() requires relationships with the `_mapping` property!',
-        );
-      }
-      const targetEntity = _mapping.targetEntity;
-      for (let targetFilterKey of _mapping.targetFilterKeys) {
-        /* type TargetFilterKey = string | string[]; */
-        if (!Array.isArray(targetFilterKey)) {
-          console.warn(
-            'WARNING: Found mapped relationship with targetFilterKey of type string. Please ensure the targetFilterKey was not intended to be of type string[]',
-          );
-          targetFilterKey = [targetFilterKey];
-        }
-        const mappingTargetEntities = entities.filter((entity) =>
-          (targetFilterKey as string[]).every(
-            (k) => targetEntity[k] === entity[k],
-          ),
-        );
-
-        if (mappingTargetEntities.length === 0) {
-          return {
-            message: () =>
-              `No target entity found for mapped relationship: ${JSON.stringify(
-                mappedRelationship,
-                null,
-                2,
-              )}`,
-            pass: false,
-          };
-        } else if (mappingTargetEntities.length > 1) {
-          return {
-            message: () =>
-              `Multiple target entities found for mapped relationship [${mappingTargetEntities.map(
-                (e) => e._key,
-              )}]; expected exactly one: ${JSON.stringify(
-                mappedRelationship,
-                null,
-                2,
-              )}`,
-            pass: false,
-          };
-        }
-      }
-    }
-    return {
-      message: () => '',
-      pass: true,
-    };
-  },
 });